refactor: separate logic into a custom hook
