+++ src/pages/MainPage.tsx
@@ -6,46 +6,31 @@ import Modal from 'components/Modal';
 import NoteItem from 'components/NoteItem';
 import MainPageImage from 'components/SvgElements/MainPageImage';
 
-import defaultNote from 'utils/constants';
-import { createNote, deleteNote, getNotes, updateNote } from 'utils/dataBase';
 import sortByDate from 'utils/functions';
-import Note from 'utils/interfaces';
+
+import useNotes from 'hooks/useNotes';
 
 function MainPage() {
-  const [defaultNotes, setDefaultNotes] = useState<Note[]>([]);
-  const [notes, setNotes] = useState<Note[]>([]);
-  const [newNote, setNewNote] = useState<Note>(defaultNote);
   const [isLoading, setLoading] = useState(false);
-  const [isModalOpen, setModalOpen] = useState(false);
-  const [editedNote, setEditedNote] = useState<Note>(defaultNote);
 
-  const getNotesFromDatabase = async () => {
-    const notesFromDataBase = await getNotes();
-    setNotes(notesFromDataBase);
-    setDefaultNotes(notesFromDataBase);
-  };
-
-  const createNewNote = async () => {
-    await createNote(newNote);
-    await getNotesFromDatabase();
-    setNewNote(defaultNote);
-  };
-
-  const removeNote = async (id: string) => {
-    await deleteNote(id);
-    await getNotesFromDatabase();
-  };
-
-  const editNote = async () => {
-    await updateNote(editedNote);
-    await getNotesFromDatabase();
-    setModalOpen(false);
-  };
+  const {
+    notes,
+    newNote,
+    setNewNote,
+    isModalOpen,
+    setModalOpen,
+    editedNote,
+    setEditedNote,
+    getNotes,
+    createNote,
+    deleteNote,
+    editNote,
+  } = useNotes();
 
   useEffect(() => {
     (async () => {
       setLoading(true);
-      await getNotesFromDatabase();
+      await getNotes();
       setLoading(false);
     })();
   }, []);
@@ -59,7 +44,7 @@ function MainPage() {
             buttonText="Create"
             note={newNote}
             setNote={setNewNote}
-            onSubmit={createNewNote}
+            onSubmit={() => createNote(newNote)}
           />
         </div>
         <div className="notes-creation__image">
@@ -76,7 +61,7 @@ function MainPage() {
               key={note.id}
               title={note.title}
               text={note.text}
-              deleteNote={() => removeNote(note.id)}
+              deleteNote={() => deleteNote(note.id)}
               editNote={() => {
                 setEditedNote(note);
                 setModalOpen(true);
